+++ include/nfx/silicon/Chip.h
@@ -1,0 +1,8 @@
+#pragma once
+
+/**
+ * \file Chip.h
+ * \brief Convenience header for all chip packages.
+ */
+
+#include "chip/cdp1xxx/CDP1852.h"